# api-docs.json
{"openapi":"3.0.1","info":{"title":"Sport Shop Orders API ","description":"Sport Shop API for orders managing with Mongo DB."},"servers":[{"url":"http://localhost:8050","description":"Generated server url"}],"security":[{"Basic Authorization":[]}],"paths":{"/api/orders":{"get":{"tags":["Orders"],"summary":"Get all orders / get orders by user ID.","description":"As Admin you can get all orders or orders owned by specified user. As User you can get all your orders.","operationId":"getOrders","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}},{"name":"customerId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK - After properly constructed request it returns orders list.","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderDto"}}}}},"400":{"description":"Bad Request - After discovering that any obligatory field is missing either in request body or in the database.","content":{"*/*":{"example":"Access denied."}}}}},"post":{"tags":["order-controller"],"operationId":"createOrder","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestOrderDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderDto"}}}}}}},"/api/orders/{orderId}":{"delete":{"tags":["order-controller"],"operationId":"deleteOrder","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}},{"name":"orderId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"RequestOrderDto":{"required":["orderItems"],"type":"object","properties":{"customerId":{"pattern":"^[a-zA-Z0-9_-]+$","type":"string"},"orderItems":{"type":"array","items":{"$ref":"#/components/schemas/RequestOrderItemDto"}}}},"RequestOrderItemDto":{"required":["productId","quantity"],"type":"object","properties":{"productId":{"type":"string"},"quantity":{"maximum":1000000,"minimum":1,"type":"integer","format":"int32"}}},"OrderDto":{"type":"object","properties":{"id":{"type":"string"},"customerId":{"type":"string"},"orderItems":{"type":"array","items":{"$ref":"#/components/schemas/OrderItemDto"}}}},"OrderItemDto":{"type":"object","properties":{"product":{"$ref":"#/components/schemas/ProductDto"},"quantity":{"type":"integer","format":"int32"}}},"ProductDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"}}}},"securitySchemes":{"Basic Authorization":{"type":"http","description":"Basic Authorization Scheme","scheme":"basic"}}}}